amazing: two curls and it works?
Smells like firmware bug, something weird in its state.

When I tried to make this use only a Perl HTTP request, it kept not
working.  I'd get HTTP streaming response, but no datagram.  If I fired
up curl in another terminal, I'd start getting datagrams.

For a while, I thought it was that curl was doing something different.
Eventually, it turned out that if I eliminated the Perl request and just
ran curl, still no good.  But if I ran curl *twice*, it worked.

To me this smells like I've found a state-management bug in the
firmware, and maybe a connection counter has gotten to be -1 instead of
0, so I need to connect twice to get it to send packets?  But who can
say.  All this touch stuff feels very janky.

For now, I'll merge this, as it sort of shows how this could work, even
though it's a bit of a crapfest.